Add optional PINENTRY_MODE_LOOPBACK for gpg>=2.1 with pygpgme>=3.1.
Note the "current" (3.0) versions of pygpgme seem to lack the
settings.

> For GnuPG >= 2.1 the pinentry mode has to be set to
> GPGME_PINENTRY_MODE_LOOPBACK to enable the passphrase callback. See
> gpgme_set_pinentry_mode.

The following repo had the appropriate hooks:
https://github.com/rshk/pygpgme
https://github.com/rshk/pygpgme/blob/6c7566b46d894ee46941f6595b4003dcc44b1104/src/pygpgme-context.c#L461..L463

For some reason this refused to work immediately, but after doing and
undoing all kinds of settings (use-agent, no-use-agent,
allow-loopback-pinentry) it started to work. Totally in the dark
what changed in the end...