gpg: Fix max() expiry date for multiple signatures
When you fix your GPG key increasing the expiry _before_ it is expired,
clean sigs will not clean up the previous signature packets. The min()
in the gpgkey module caused the oldest signatures to be used, showing an
older expiry than the latest and greatest.

Now the signatures are grouped by packet-to-be-signed and the max()
expiry is taken in those. Then the min() expiry of all packets is taken.